// backend/src/infrastructure/external/connectors/walmart/walmart.types.ts
export interface Order {
  purchaseOrderId: string;
  customerOrderId: string;
  customerEmailId: string;
  customerRfc: string;
  orderDate: number;
  orderSummary: OrderSummary;
  shippingInfo: ShippingInfo;
  orderLines: {
    orderLine: OrderLine[];
  };
}

export interface OrderSummary {
  totalAmount: Money;
  orderSubTotals: {
    subTotalType: string;
    totalAmount: Money;
  }[];
}

export interface Money {
  amount: number;
  currency: string;
}

export interface ShippingInfo {
  phone: string;
  estimatedDeliveryDate: number;
  estimatedShipDate: number;
  methodCode: string;
  postalAddress: PostalAddress;
}

export interface PostalAddress {
  name: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
}

export interface OrderLine {
  lineNumber: string;
  item: {
    productName: string;
    sku: string;
  };
  charges: {
    charge: {
      chargeType: string;
      chargeName: string;
      chargeAmount: Money;
      tax: {
        taxAmount: Money;
      };
    }[];
  };
  orderLineQuantity: {
    unitOfMeasurement: string;
    amount: number;
  };
  orderLineStatuses: {
    orderLineStatus: {
      status: string;
      statusQuantity: {
        unitOfMeasurement: string;
        amount: number;
      };
      trackingInfo: {
        carrierName: {
          otherCarrier: string;
          carrier: string;
        };
        trackingNumber: string;
        methodCode: string;
        shipmentNo: string;
      };
      cancellationReason: string;
    }[];
  };
  fulfillment: {
    fulfillmentOption: string;
    shipMethod: string;
    storeId: string;
    offerId: string;
    predictedShipNodeId: string;
    predictedShipNodeName: string;
  };
  statusDate: number;
}